refactor: move project form to widget
