+++ crates/ui/src/widgets.rs
@@ -1,5 +1,6 @@
 //! This module contains various reusable widgets that can be shared across various UI layouts.
 
+pub mod create_project_form;
 pub mod notifications;
 pub mod status_bar;
 pub mod toolbar;
