+++ locales/en-GB/widgets.ftl
@@ -1,0 +1,3 @@
+# Format: {domain}-{component}-{element}-{variant}
+
+widgets-create_project_form-title = Create new project
